new resolving itf & minor changes

+++ src/tcp.h
@@ -7,7 +7,7 @@
 
 class TcpClient : public IOWriteable, public error_handler {
 public:
-    virtual void init(const std::string& host, int port, IOLoop* loop) = 0;
+    virtual void init(const std::string& host, uint16_t port, IOLoop* loop) = 0;
     virtual void on_read(OnReadFunc func) = 0;
     virtual void on_connect(OnEventFunc func) = 0;
     virtual void on_close(OnEventFunc func) = 0;
@@ -24,7 +24,7 @@ public:
 class TcpServer : public error_handler {
 public:
     using OnConnectFunc = std::function<void(std::unique_ptr<TcpSocket>&, sockaddr*, socklen_t)>;
-    virtual void init(int port, IOLoop* loop, const std::string& host_or_interface="") = 0;
+    virtual void init(uint16_t port, IOLoop* loop, const std::string& host_or_interface="") = 0;
     virtual void on_connect(OnConnectFunc func) = 0;
     virtual void on_close(OnEventFunc func) = 0;
     static std::unique_ptr<TcpServer> create(const std::string& name);
